+++ src/Symfony/Component/Serializer/Encoder/JsonEncode.php
@@ -26,18 +26,9 @@ class JsonEncode implements EncoderInterface
         self::OPTIONS => 0,
     ];
 
-    /**
-     * @param array $defaultContext
-     */
-    public function __construct($defaultContext = [])
+    public function __construct(array $defaultContext = [])
     {
-        if (!\is_array($defaultContext)) {
-            @trigger_error(sprintf('Passing an integer as first parameter of the "%s()" method is deprecated since Symfony 4.2, use the "json_encode_options" key of the context instead.', __METHOD__), E_USER_DEPRECATED);
-
-            $this->defaultContext[self::OPTIONS] = (int) $defaultContext;
-        } else {
-            $this->defaultContext = array_merge($this->defaultContext, $defaultContext);
-        }
+        $this->defaultContext = array_merge($this->defaultContext, $defaultContext);
     }
 
     /**