+++ src/Symfony/Component/Serializer/Normalizer/DataUriNormalizer.php
@@ -12,8 +12,6 @@
 namespace Symfony\Component\Serializer\Normalizer;
 
 use Symfony\Component\HttpFoundation\File\File;
-use Symfony\Component\HttpFoundation\File\MimeType\MimeTypeGuesser;
-use Symfony\Component\HttpFoundation\File\MimeType\MimeTypeGuesserInterface as DeprecatedMimeTypeGuesserInterface;
 use Symfony\Component\Mime\MimeTypeGuesserInterface;
 use Symfony\Component\Mime\MimeTypes;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
@@ -38,22 +36,10 @@ class DataUriNormalizer implements NormalizerInterface, DenormalizerInterface, C
      */
     private $mimeTypeGuesser;
 
-    /**
-     * @param MimeTypeGuesserInterface|null $mimeTypeGuesser
-     */
-    public function __construct($mimeTypeGuesser = null)
+    public function __construct(MimeTypeGuesserInterface $mimeTypeGuesser = null)
     {
-        if ($mimeTypeGuesser instanceof DeprecatedMimeTypeGuesserInterface) {
-            @trigger_error(sprintf('Passing a %s to "%s()" is deprecated since Symfony 4.3, pass a "%s" instead.', DeprecatedMimeTypeGuesserInterface::class, __METHOD__, MimeTypeGuesserInterface::class), E_USER_DEPRECATED);
-        } elseif (null === $mimeTypeGuesser) {
-            if (class_exists(MimeTypes::class)) {
-                $mimeTypeGuesser = MimeTypes::getDefault();
-            } elseif (class_exists(MimeTypeGuesser::class)) {
-                @trigger_error(sprintf('Passing null to "%s()" to use a default MIME type guesser without Symfony Mime installed is deprecated since Symfony 4.3. Try running "composer require symfony/mime".', __METHOD__), E_USER_DEPRECATED);
-                $mimeTypeGuesser = MimeTypeGuesser::getInstance();
-            }
-        } elseif (!$mimeTypeGuesser instanceof MimeTypes) {
-            throw new \TypeError(sprintf('Argument 1 passed to "%s()" must be an instance of "%s" or null, %s given.', __METHOD__, MimeTypes::class, \is_object($mimeTypeGuesser) ? \get_class($mimeTypeGuesser) : \gettype($mimeTypeGuesser)));
+        if (!$mimeTypeGuesser && class_exists(MimeTypes::class)) {
+            $mimeTypeGuesser = MimeTypes::getDefault();
         }
 
         $this->mimeTypeGuesser = $mimeTypeGuesser;
@@ -112,6 +98,10 @@ class DataUriNormalizer implements NormalizerInterface, DenormalizerInterface, C
         try {
             switch ($class) {
                 case 'Symfony\Component\HttpFoundation\File\File':
+                    if (!class_exists(File::class)) {
+                        throw new InvalidArgumentException(sprintf('Cannot denormalize to a "%s" without the HttpFoundation component installed. Try running "composer require symfony/http-foundation".', File::class));
+                    }
+
                     return new File($data, false);
 
                 case 'SplFileObject':
@@ -152,10 +142,6 @@ class DataUriNormalizer implements NormalizerInterface, DenormalizerInterface, C
     {
         if ($object instanceof File) {
             return $object->getMimeType();
-        }
-
-        if ($this->mimeTypeGuesser instanceof DeprecatedMimeTypeGuesserInterface && $mimeType = $this->mimeTypeGuesser->guess($object->getPathname())) {
-            return $mimeType;
         }
 
         if ($this->mimeTypeGuesser && $mimeType = $this->mimeTypeGuesser->guessMimeType($object->getPathname())) {